+++ src/pages/index.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { useHead } from '@vueuse/head'
-
+import Title from '~/components/pcs/Title.vue'
+import Alert from '~/components/pcs/Alert.vue'
 useHead({
   title: 'Index Title',
   meta: [
@@ -11,7 +12,8 @@ useHead({
 
 <template>
   <div>
-    <h1>Index</h1>
+    <Title name="Index" />
+    <Alert />
   </div>
 </template>