+++ src/pages/test.vue
@@ -1,23 +1,25 @@
 <script setup lang="ts">
 import { useHead } from '@vueuse/head'
 import Title from '~/components/pcs/Title.vue'
-
+import Alert from '~/components/pcs/Alert.vue'
 useHead({
-  title: '404 Not Found',
+  title: 'Test Title',
   meta: [
-    { name: 'description', content: 'Not Found' },
+    { name: 'description', content: 'Test Description' },
   ],
 })
 </script>
 
 <template>
   <div>
-    <Title name="404" />
+    <Title name="Test" />
+    # not working
+    <Alert />
   </div>
 </template>
 
 <route lang="yaml">
 meta:
-  layout: error
+  layout: content
   requiresAuth: false
 </route>